Upgrade Login with MVC -- wait for Register
