+++ app/src/main/java/com/example/touristo/Controlers/Auth/Login.java
@@ -1,108 +1,50 @@
 package com.example.touristo.Controlers.Auth;
 
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
-import android.os.Bundle;
-import android.text.TextUtils;
 import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
 import android.widget.ProgressBar;
-import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+
 import com.example.touristo.MainActivity;
-import com.example.touristo.R;
+import com.example.touristo.Views.LoginActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
-public class Login extends AppCompatActivity {
-    EditText login_email;
-    EditText login_password;
-    Button login_button;
-    TextView signupRedirectText;
+public class Login {
+
     FirebaseAuth mAuth;
-    ProgressBar progressBar;
+    LoginActivity loginActivity;
 
-    @Override
-    public void onStart() {
-        super.onStart();
-        FirebaseUser currentUser = mAuth.getCurrentUser();
-
-        if(currentUser != null){
-            Intent intent = new Intent(getApplicationContext() , MainActivity.class);
-            startActivity(intent);
-            finish();
-        }
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.login);
-
+    public Login(LoginActivity loginActivity){
         mAuth = FirebaseAuth.getInstance();
-
-        login_email = findViewById(R.id.login_email);
-        login_password = findViewById(R.id.login_password);
-        login_button = findViewById(R.id.login_button);
-        signupRedirectText = findViewById(R.id.signupRedirectText);
-        progressBar = findViewById(R.id.progressBar);
-
-        //moving from login activity to the signup activity
-        signupRedirectText.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                Intent intent = new Intent(Login.this, Register.class );
-                startActivity(intent);
-                finish();
-            }
-        });
-
-        //login user
-        login_button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                progressBar.setVisibility(View.VISIBLE);
-                String email , password ;
-                email = login_email.getText().toString();
-                password = login_password.getText().toString();
-
-                if (TextUtils.isEmpty(email))
-                {
-                    Toast.makeText(Login.this, "Enter an email!!!", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                if (TextUtils.isEmpty(password))
-                {
-                    Toast.makeText(Login.this, "Enter a password!!!", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-
-                mAuth.signInWithEmailAndPassword(email, password)
-                        .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-                            @Override
-                            public void onComplete(@NonNull Task<AuthResult> task) {
-                                progressBar.setVisibility(View.GONE);
-                                if (task.isSuccessful()) {
-                                    Toast.makeText(Login.this, "Authentication succeed.",Toast.LENGTH_SHORT).show();
-                                    Intent intent = new Intent( getApplicationContext(), MainActivity.class );
-                                    startActivity(intent);
-                                    finish();
-                                } else {
-                                    // If sign in fails, display a message to the user.
-                                    Toast.makeText(Login.this, "Authentication failed.",
-                                            Toast.LENGTH_SHORT).show();
-                                }
-                            }
-                        });
-            }
-        });
+        this.loginActivity = loginActivity;
     }
 
+    public FirebaseUser getCurrentUser(){
+        return mAuth.getCurrentUser();
+    }
+
+    public void signUp(String email, String password, ProgressBar progressBar){
+        mAuth.signInWithEmailAndPassword(email, password)
+                .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        progressBar.setVisibility(View.GONE);
+                        if (task.isSuccessful()) {
+                            Intent intent = new Intent( loginActivity.getApplicationContext(), MainActivity.class );
+                            loginActivity.startActivity(intent);
+                            loginActivity.finish();
+                        } else {
+                            // If sign in fails, display a message to the user.
+                            Toast.makeText(loginActivity, "Authentication failed.",
+                                    Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
+    }
 }
