+++ app/src/main/java/com/example/touristo/MainActivity.java
@@ -10,6 +10,7 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.example.touristo.Controlers.Auth.Login;
+import com.example.touristo.Views.LoginActivity;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
@@ -31,7 +32,7 @@ public class MainActivity extends AppCompatActivity {
         mAuth = FirebaseAuth.getInstance();
         user = mAuth.getCurrentUser();
         if(user == null){
-            Intent intent = new Intent(getApplicationContext() , Login.class);
+            Intent intent = new Intent(getApplicationContext() , LoginActivity.class);
             startActivity(intent);
             finish();
         }
@@ -43,7 +44,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
              FirebaseAuth.getInstance().signOut();
-                Intent intent = new Intent(getApplicationContext() , Login.class);
+                Intent intent = new Intent(getApplicationContext() , LoginActivity.class);
                 startActivity(intent);
                 finish();
             }
