--- app/src/main/java/com/example/touristo/Views/Login.java
@@ -1,4 +1,0 @@
-package com.example.touristo.Views;
-
-public class Login {
-}
