+++ app/src/main/java/com/example/touristo/Views/RegisterActivity.java
@@ -1,4 +1,5 @@
 package com.example.touristo.Views;
 
-public class Register {
+public class RegisterActivity {
+
 }
